requirements.txt
beanie
fastapi
pymongo
motor
pydantic
python-multipart
python-dotenv
pydantic-settings
uvicorn